+++ CHANGELOG.md
@@ -8,7 +8,9 @@ v0.11 includes:
 * Added isGenericType function to templates.
 * Corrected logic in getApiItems.
 * Moved global css/js files to be after local css/js files.
-* Updated createAnchorToType to include a parameter to assign it a class.
+* Updated createAnchorToType to include a parameter for multiple options, like:
+  * Assigning classes to the anchor tag.
+  * Setting the display name to full name instead of just the name of the type.
 
 v0.10 includes:
 * Updated getApiItems to include ParameterInfo and GenericParameterInfo details for undocumented items to appear.